Miniter/requirements.txt
sqlalchemy
mysql-connector-python
PyJWT==1.4.0
flask-cors